requirements.txt
asgiref==3.2.0
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
autopep8==1.4.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.40.0
django-countries==5.5
django-crispy-forms==1.7.2
django-tinymce4-lite==1.7.5
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
jsmin==2.2.2
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==6.1.0
psycopg2==2.8.3
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.2.4
stripe==2.37.1
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.6.1
virtualenvwrapper-win==1.2.5
wrapt==1.11.2
zope.interface==4.6.0
